+++ src/main/java/com/trax/dao/user/UserDAO.java
@@ -15,9 +15,9 @@ public interface UserDAO {
 
     public void addUser(User user);
     public void updateUser(User user);
-    public User getUser(int id);
+    public User getUser(Long id);
     public User getUser(String username);
-    public void deleteUser(int id);
+    public void deleteUser(Long id);
     public List getUsers();
 
 }
