+++ src/main/java/com/trax/services/contact/ContactService.java
@@ -15,8 +15,8 @@ public interface ContactService {
 
     public void addContact(Contact contact);
     public void updateContact(Contact contact);
-    public Contact getContact(int id);
-    public void deleteContact(int id);
+    public Contact getContact(Long id);
+    public void deleteContact(Long id);
     public List<Contact> getContacts();
 
 }
